Feat : Server 삭제 권한 관련 에러처리 추가
- 삭제하려는 사용자가 Server의 owner가 아닐 경우 에러를 발생시키도록 변경하였습니다.
- 상태코드 및 throw하는 exception을 더 명확하게 처리하였습니다.

+++ backend/src/server/server.service.ts
@@ -1,4 +1,10 @@
-import { forwardRef, Inject, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ForbiddenException,
+  forwardRef,
+  Inject,
+  Injectable,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { User } from '../user/user.entity';
@@ -6,6 +12,7 @@ import { Server } from './server.entity';
 import RequestServerDto from './dto/RequestServerDto';
 import { UserServerService } from '../user-server/user-server.service';
 import { ServerRepository } from './server.repository';
+import { DeleteResult } from 'typeorm';
 
 @Injectable()
 export class ServerService {
@@ -47,8 +54,17 @@ export class ServerService {
     await this.serverRepository.update(id, server);
   }
 
-  async deleteServer(id: number): Promise<void> {
-    await this.serverRepository.delete({ id: id });
+  async deleteServer(id: number, user: User) {
+    const server = await this.serverRepository.findOneWithOwner(id);
+
+    if (server.owner.id !== user.id) {
+      throw new ForbiddenException('삭제 권한이 없습니다.');
+    }
+    if (!server) {
+      throw new BadRequestException('해당 서버가 존재하지 않습니다.');
+    }
+
+    return this.serverRepository.delete({ id: id });
   }
 
   findOneWithUsers(serverId: number): Promise<Server> {
